Added new methods for metaheuristic.

+++ src/basenet/metaheuristic/basis/basic_evolutive/elitist_selection.py
@@ -10,7 +10,7 @@ import tensorflow as tf
 
 # -----------------------------------------------------------
 def elitist_selection(new_individuals: tf.Tensor, population: tf.Tensor) -> tf.Tensor:
-    return tf.concat([population[0:-len(new_individuals)], new_individuals], 0)
+    return tf.cast(tf.concat([population[0:-len(new_individuals)], new_individuals], 0), dtype=tf.float32)
 # - x - x - x - x - x - x - x - x - x - x - x - x - x - x - #
 #                        END OF FILE                        #
 # - x - x - x - x - x - x - x - x - x - x - x - x - x - x - #
